+++ src/Models/HeaderConfig.php
@@ -39,65 +39,9 @@ class HeaderConfig extends BaseServiceConfigModel
     public $incrementing = true;
 
     /**
-     * {@inheritdoc}
+     * @var bool
      */
-    public static function getConfig($id, $protect = true)
-    {
-        $params = static::whereServiceId($id);
-
-        if (!empty($params)) {
-            return $params->toArray();
-        } else {
-            return [];
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function validateConfig($config, $create = true)
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function setConfig($id, $config)
-    {
-        static::whereServiceId($id)->delete();
-        if (!empty($config)) {
-            foreach ($config as $header) {
-                //Making sure service_id is the first item in the config.
-                //This way service_id will be set first and is available
-                //for use right away. This helps setting an auto-generated
-                //field that may depend on parent data. See OAuthConfig->setAttribute.
-                $header = array_reverse($header, true);
-                $header['service_id'] = $id;
-                $header = array_reverse($header, true);
-                static::create($header);
-            }
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function getConfigSchema()
-    {
-        $schema =
-            [
-                'name'        => 'headers',
-                'label'       => 'Headers',
-                'description' => 'Supply additional headers to pass to the remote service.',
-                'type'        => 'array',
-                'required'    => false,
-                'allow_null'  => true
-            ];
-        $schema['items'] = parent::getConfigSchema();
-
-        return $schema;
-    }
+    public static $alwaysNewOnSet = true;
 
     /**
      * @param array $schema