+++ src/Models/ParameterConfig.php
@@ -7,6 +7,7 @@ use DreamFactory\Core\Models\BaseServiceConfigModel;
  * ParameterConfig
  *
  * @property integer $id
+ * @property integer $service_id
  * @property string  $name
  * @property string  $value
  * @property boolean $exclude
@@ -43,65 +44,9 @@ class ParameterConfig extends BaseServiceConfigModel
     public $incrementing = true;
 
     /**
-     * {@inheritdoc}
+     * @var bool
      */
-    public static function getConfig($id, $protect = true)
-    {
-        $params = static::whereServiceId($id);
-
-        if (!empty($params)) {
-            return $params->toArray();
-        } else {
-            return [];
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function validateConfig($config, $create = true)
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function setConfig($id, $config)
-    {
-        static::whereServiceId($id)->delete();
-        if (!empty($config)) {
-            foreach ($config as $param) {
-                //Making sure service_id is the first item in the config.
-                //This way service_id will be set first and is available
-                //for use right away. This helps setting an auto-generated
-                //field that may depend on parent data. See OAuthConfig->setAttribute.
-                $param = array_reverse($param, true);
-                $param['service_id'] = $id;
-                $param = array_reverse($param, true);
-                static::create($param);
-            }
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function getConfigSchema()
-    {
-        $schema =
-            [
-                'name'        => 'parameters',
-                'label'       => 'Parameters',
-                'description' => 'Supply additional parameters to pass to the remote service, or exclude parameters passed from client.',
-                'type'        => 'array',
-                'required'    => false,
-                'allow_null'  => true
-            ];
-        $schema['items'] = parent::getConfigSchema();
-
-        return $schema;
-    }
+    public static $alwaysNewOnSet = true;
 
     /**
      * @param array $schema