+++ src/Services/RemoteWeb.php
@@ -4,11 +4,16 @@ use Config;
 use DreamFactory\Core\Components\Cacheable;
 use DreamFactory\Core\Contracts\CachedInterface;
 use DreamFactory\Core\Contracts\HttpStatusCodeInterface;
+use DreamFactory\Core\Contracts\ServiceResponseInterface;
 use DreamFactory\Core\Enums\ApiOptions;
 use DreamFactory\Core\Enums\HttpStatusCodes;
+use DreamFactory\Core\Enums\ServiceTypeGroups;
 use DreamFactory\Core\Enums\VerbsMask;
 use DreamFactory\Core\Exceptions\InternalServerErrorException;
 use DreamFactory\Core\Exceptions\RestException;
+use DreamFactory\Core\Http\Controllers\StatusController;
+use DreamFactory\Core\Models\Service;
+use DreamFactory\Core\Resources\System\ServiceType;
 use DreamFactory\Core\Services\BaseRestService;
 use DreamFactory\Core\Utility\ResourcesWrapper;
 use DreamFactory\Core\Utility\ResponseFactory;
@@ -16,6 +21,7 @@ use DreamFactory\Core\Utility\Session;
 use DreamFactory\Library\Utility\Curl;
 use DreamFactory\Library\Utility\Enums\Verbs;
 use DreamFactory\Library\Utility\Scalar;
+use Illuminate\Contracts\Support\Arrayable;
 use Log;
 
 class RemoteWeb extends BaseRestService implements CachedInterface
@@ -30,6 +36,12 @@ class RemoteWeb extends BaseRestService implements CachedInterface
      * @var string
      */
     protected $baseUrl;
+
+    /**
+     * @var boolean
+     */
+    protected $replaceLinks = false;
+
     /**
      * @var array
      */
@@ -72,10 +84,24 @@ class RemoteWeb extends BaseRestService implements CachedInterface
         $this->autoDispatch = false;
 
         $config = (array)array_get($settings, 'config', []);
+        $this->apiDoc = (array)array_get($settings, 'doc');
         $this->baseUrl = array_get($config, 'base_url');
+        $this->replaceLinks = array_get($config, 'replace_link', false);
         // Validate url setup
         if (empty($this->baseUrl)) {
-            throw new \InvalidArgumentException('Remote Web Service base url can not be empty.');
+            // fancy trick to grab the base url from swagger
+            if (!empty($this->apiDoc) && !empty($host = array_get($this->apiDoc, 'host'))) {
+                if (!empty($protocol = array_get($this->apiDoc, 'schemes'))) {
+                    $protocol = (is_array($protocol) ? current($protocol) : $protocol);
+                } else {
+                    $protocol = 'http';
+                }
+                $basePath = array_get($this->apiDoc, 'basePath', '');
+                $this->baseUrl = $protocol . '://' . $host . $basePath;
+            }
+            if (empty($this->baseUrl)) {
+                throw new \InvalidArgumentException('Remote Web Service base url can not be empty.');
+            }
         }
 
         $this->options = (array)array_get($config, 'options');
@@ -87,7 +113,6 @@ class RemoteWeb extends BaseRestService implements CachedInterface
         $this->cacheTTL = intval(array_get($config, 'cache_ttl', Config::get('df.default_cache_ttl')));
         $this->cachePrefix = 'service_' . $this->id . ':';
 
-        $this->apiDoc = (array)array_get($settings, 'doc');
         $this->implementsAccessList = boolval(array_get($config, 'implements_access_list', false));
     }
 
@@ -106,7 +131,7 @@ class RemoteWeb extends BaseRestService implements CachedInterface
         $value,
         $add_to_query = true,
         $add_to_key = true
-    ){
+    ) {
         if (is_array($value)) {
             foreach ($value as $sub => $subValue) {
                 static::parseArrayParameter($query,
@@ -157,6 +182,7 @@ class RemoteWeb extends BaseRestService implements CachedInterface
      * @param string $action
      * @param string $query
      * @param string $cache_key
+     * @param array  $requestParams
      *
      * @return void
      */
@@ -314,7 +340,7 @@ class RemoteWeb extends BaseRestService implements CachedInterface
 
     /**
      * @throws \DreamFactory\Core\Exceptions\RestException
-     * @return bool
+     * @return ServiceResponseInterface
      */
     protected function processRequest()
     {
@@ -402,6 +428,7 @@ class RemoteWeb extends BaseRestService implements CachedInterface
         }
 
         $contentType = Curl::getInfo('content_type');
+        $result = $this->fixLinks($result);
         $response = ResponseFactory::create($result, $contentType, $status);
         $response->setHeaders($resultHeaders);
 
@@ -414,6 +441,46 @@ class RemoteWeb extends BaseRestService implements CachedInterface
         }
 
         return $response;
+    }
+
+    /**
+     * Replaces any hyper links referencing any DF remote web service's base url
+     * with the link to the corresponding DF service itself to ensure DF rws works
+     * as the gateway to all remote service calls.
+     *
+     * @param $result
+     *
+     * @return mixed
+     */
+    protected function fixLinks($result)
+    {
+        if (false === $this->replaceLinks) {
+            return $result;
+        }
+
+        $isArray = false;
+        $string = '';
+        if (is_array($result)) {
+            $string = json_encode($result, JSON_UNESCAPED_SLASHES);
+            $isArray = true;
+        } elseif (is_string($result)) {
+            $string = $result;
+        }
+
+        $myUri = StatusController::getURI($_SERVER);
+
+        $allRws = Service::whereType('rws')->whereIsActive(1)->get()->all();
+        if (!empty($allRws)) {
+            /** @var Service $rws */
+            foreach ($allRws as $rws) {
+                $config = $rws->getConfigAttribute();
+                $baseUrl = trim(array_get($config, 'base_url'), '/');
+                $dfUrl = trim($myUri, '/') . '/api/v2/' . $rws->name;
+                $string = str_replace($baseUrl, $dfUrl, $string);
+            }
+        }
+
+        return ($isArray) ? json_decode($string, true) : $string;
     }
 
     protected function getEventName()